Transformations

Classical Orbital Elements (coe)

AstroRepresentations.convert6_cart_to_coe — Function

convert6_cart_to_coe(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert cartesian state representation into classical orbital elements.

Inputs

sv – state vector – L, T

μ – center gravitational parameter – L³/T²

Output

Classical orbital elements representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

AstroRepresentations.convert6_coe_to_cart — Function

convert6_coe_to_cart(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert classical orbital elements state vector to cartesian state.

Inputs

sv – Keplerian elements – L, rad

μ – Center's gravitational parameter – L³/T²

Output

Cartesian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

AstroRepresentations.∂convert6_cart_to_coe — Function

∂convert6_cart_to_coe(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert cartesian state representation into classical orbital elements. Compute also the full jacobian of the elements wrt the cartesian state.

Inputs

sv – state vector – L, T

μ – center gravitational parameter – L³/T²

Output

Classical Orbital Elements representation of the state as a SVector{6} and its jacobian as a SMatrix{6, 6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

Pasquale, A. - Multiple Shooting Optimiser (MSO). Technical Note 0001, 2022.

AstroRepresentations.∂convert6_coe_to_cart — Function

∂convert6_coe_to_cart(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert classical orbital elements state vector to cartesian state. Compute also the full jacobian of the cartesian states wrt the elements.

Inputs

sv – cartesian state representation – L, rad

μ – Center's gravitational parameter – L³/T²

Output

Cartesian representation of the state as a SVector{6} and its jacobian as a SMatrix{6, 6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications. Vol. 12. Springer Science & Business Media, 2001.

Pasquale, A. - Multiple Shooting Optimiser (MSO). Technical Note 0001, 2022.

Equinoctial Orbital Elements (eoe)

AstroRepresentations.convert6_cart_to_equi — Function

convert6_cart_to_equi(sv::AbstractVector{<:Number}, μ::Number, [args]...)

Convert cartesian state vector to Equinoctial keplerian elements.

Inputs

sv – state vector – L, T

μ – center gravitational parameter – L³/T²

Output

Equinoctial representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

Cefola - DOI: 10.2514/6.1972-937

AstroRepresentations.convert6_equi_to_cart — Function

convert6_equi_to_cart(equi::AbstractVector{<:Number}, μ::Number, [args]...)

Convert equinoctial state elements to cartesian state.

Inputs

equi – Equinoctial elements – L, rad

μ – center gravitational parameter – L³/T²

Output

Cartesian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

Cefola - DOI: 10.2514/6.1972-937

AstroRepresentations.convert6_coe_to_equi — Function

convert6_coe_to_equi(coe::AbstractVector{<:Number}, [args]...)

Convert classical orbital elements state vector to Equinoctial keplerian elements.

Inputs

coe – Keplerian elements – L, rad

Output

Equinoctial representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

AstroRepresentations.convert6_equi_to_coe — Function

convert6_equi_to_coe(coe::AbstractVector{<:Number}, [args]...)

Convert Equinoctial keplerian elements to classical orbital elements state vector .

Inputs

equi – Equinoctial elements – L, rad

Output

Keplerian representation of the state as a SVector{6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.

AstroRepresentations.∂convert6_coe_to_equi — Function

∂convert6_coe_to_equi(coe::AbstractVector{<:Number}, [args]...)

Convert classical orbital elements state vector to Equinoctial keplerian elements. Compute also the full jacobian of the equinoctial elemenents wrt the classical orbital elements.

Inputs

coe – Keplerian elements – L, rad

Output

Equinoctial representation of the state as a SVector{6} and its jacobian as a SMatrix{6, 6}.

References

Vallado, David A. - Fundamentals of astrodynamics and applications, 2013.